Array에 updateStockLabels()을 extension하여 공통 코드를 추출했습니다.
